shl_genai_recommender/requirements.txt
fastapi
uvicorn
openai
faiss-cpu
pandas
numpy
python-dotenv